Implement the add task opertaion
